requirements.txt
feedparser
apscheduler
Pillow
flask
jsmin
slack
slackclient